Add NullBuyingPowerModel
Provides an implementation of IBuyingPowerModel that essentially
disables margin checks and ensures all orders are not failed due
to insufficient buying power. Technically, orders can still fail,
but it requires the entire portfolio to be underwater which would
be exceedingly hard to do if this model is cofigured onall secrites.